refactor weight question
